Pass size_t value into base::span constructors.
base::span will shortly require that length is size_t.

CFDataGetLength returns a CFIndex which is a signed int.

This CL adds checked_casts for this type to size_t so
that a potentially negative length is never passed to
base::span. This prevents compile errors in a future
CL.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4215918
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101044}

+++ device/fido/mac/credential_store.mm
@@ -17,6 +17,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/mac/mac_logging.h"
 #include "base/mac/scoped_cftyperef.h"
 #include "base/metrics/histogram_functions.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/strings/sys_string_conversions.h"
 #include "components/device_event_log/device_event_log.h"
 #include "crypto/random.h"
@@ -436,9 +437,9 @@ bool TouchIdCredentialStore::DeleteCredentialsSync(
       DLOG(ERROR) << "missing application label";
       continue;
     }
-    if (!DeleteCredentialById(
-            base::make_span(CFDataGetBytePtr(credential_id_data),
-                            CFDataGetLength(credential_id_data)))) {
+    if (!DeleteCredentialById(base::make_span(
+            CFDataGetBytePtr(credential_id_data),
+            base::checked_cast<size_t>(CFDataGetLength(credential_id_data))))) {
       // Indicate failure, but keep deleting remaining items.
       result = false;
     }
