+++ net/cert/internal/trust_store_mac.cc
@@ -17,6 +17,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/mac/mac_logging.h"
 #include "base/metrics/histogram_functions.h"
 #include "base/no_destructor.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/strings/strcat.h"
 #include "base/synchronization/lock.h"
 #include "base/timer/elapsed_timer.h"
@@ -403,7 +404,8 @@ class TrustDomainCacheFullCerts {
         continue;
       }
       auto buffer = x509_util::CreateCryptoBuffer(base::make_span(
-          CFDataGetBytePtr(der_data.get()), CFDataGetLength(der_data.get())));
+          CFDataGetBytePtr(der_data.get()),
+          base::checked_cast<size_t>(CFDataGetLength(der_data.get()))));
       CertErrors errors;
       ParseCertificateOptions options;
       options.allow_invalid_serial_numbers = true;
@@ -831,7 +833,8 @@ class TrustStoreMac::TrustImplDomainCacheFullCerts
         continue;
       }
       auto buffer = x509_util::CreateCryptoBuffer(base::make_span(
-          CFDataGetBytePtr(der_data.get()), CFDataGetLength(der_data.get())));
+          CFDataGetBytePtr(der_data.get()),
+          base::checked_cast<size_t>(CFDataGetLength(der_data.get()))));
       CertErrors errors;
       ParseCertificateOptions options;
       options.allow_invalid_serial_numbers = true;
@@ -1003,7 +1006,8 @@ class TrustStoreMac::TrustImplKeychainCacheFullCerts
         continue;
       }
       auto buffer = x509_util::CreateCryptoBuffer(base::make_span(
-          CFDataGetBytePtr(der_data.get()), CFDataGetLength(der_data.get())));
+          CFDataGetBytePtr(der_data.get()),
+          base::checked_cast<size_t>(CFDataGetLength(der_data.get()))));
       CertErrors errors;
       ParseCertificateOptions options;
       options.allow_invalid_serial_numbers = true;
@@ -1237,9 +1241,10 @@ TrustStoreMac::FindMatchingCertificatesForMacNormalizedSubject(
       LOG(ERROR) << "SecCertificateCopyData error";
       continue;
     }
-    matching_cert_buffers.push_back(x509_util::CreateCryptoBuffer(
-        base::make_span(CFDataGetBytePtr(der_data.get()),
-                        CFDataGetLength(der_data.get()))));
+    matching_cert_buffers.push_back(
+        x509_util::CreateCryptoBuffer(base::make_span(
+            CFDataGetBytePtr(der_data.get()),
+            base::checked_cast<size_t>(CFDataGetLength(der_data.get())))));
   }
   return matching_cert_buffers;
 }
