+++ net/cert/x509_util_apple.cc
@@ -11,6 +11,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "base/check_op.h"
 #include "base/logging.h"
+#include "base/numerics/safe_conversions.h"
 #include "build/build_config.h"
 #include "net/cert/x509_certificate.h"
 #include "third_party/boringssl/src/include/openssl/pool.h"
@@ -30,7 +31,8 @@ bssl::UniquePtr<CRYPTO_BUFFER> CertBufferFromSecCertificate(
     return nullptr;
   }
   return X509Certificate::CreateCertBufferFromBytes(
-      base::make_span(CFDataGetBytePtr(der_data), CFDataGetLength(der_data)));
+      base::make_span(CFDataGetBytePtr(der_data),
+                      base::checked_cast<size_t>(CFDataGetLength(der_data))));
 }
 
 }  // namespace
